Moved some bits around
