+++ workflows/src/main/kotlin/com/r3/corda/lib/tokens/workflows/utilities/TokenUtilities.kt
@@ -1,5 +1,6 @@
 @file:JvmName("TokenUtilities")
-package com.r3.corda.lib.tokens.workflows.flows.shell
+
+package com.r3.corda.lib.tokens.workflows.utilities
 
 import com.r3.corda.lib.tokens.contracts.states.NonFungibleToken
 import com.r3.corda.lib.tokens.contracts.types.IssuedTokenType
@@ -10,6 +11,7 @@ import net.corda.core.identity.AbstractParty
 /**
  * Creates a [NonFungibleToken] from an [IssuedTokenType].
  * E.g. IssuedTokenType<TokenType> -> NonFungibleToken<TokenType>.
+ * This function must exist outside of the contracts module as creating a unique identifier is non deterministic.
  */
 infix fun <T : TokenType> IssuedTokenType<T>.heldBy(owner: AbstractParty): NonFungibleToken<T> = _heldBy(owner)
